fix(util): make statfs block-size math portable across darwin/linux
syscall.Statfs_t.Bsize is uint32 on darwin but int64 on linux, so
GetFilesystemStats failed to compile on macOS. Convert once to int64
so the package builds and tests run on both platforms.

+++ pkg/util/mount.go
@@ -226,7 +226,7 @@ func GetFilesystemStats(path string) (*FilesystemStats, error) {
 		return nil, fmt.Errorf("statfs failed: %w", err)
 	}
 
-	blockSize := stat.Bsize
+	blockSize := int64(stat.Bsize)
 
 	return &FilesystemStats{
 		TotalBytes:      int64(stat.Blocks) * blockSize,
